UC-2200: * Changed time parameters from int to unsigned long. * Added expireTime on all queries for unexpired records * Added 500 response code for successful registrations that produced no contacts from regdb * Adjust subscription database function calls to prefer primary * Added shardId in subscription database expiration removal * Improved logging

+++ sipXcommserverLib/src/persist/SipPersistentSubscriptionMgr.cpp
@@ -400,7 +400,7 @@ void SipPersistentSubscriptionMgr::updateVersion(SipMessage& notifyRequest,
    UtlString from;
    UtlString callId;
    UtlString eventHeader, eventId;
-   int now;
+   unsigned long now;
 
    // Note that the "to" and "from" fields of the subscription table
    // are as those URIs appear in the SUBSCRIBE message, which is
@@ -409,7 +409,7 @@ void SipPersistentSubscriptionMgr::updateVersion(SipMessage& notifyRequest,
    notifyRequest.getFromField(&to);
    notifyRequest.getCallIdField(&callId);
    notifyRequest.getEventFieldParts(&eventHeader, &eventId);
-   now = (int) OsDateTime::getSecsSinceEpoch();
+   now = OsDateTime::getSecsSinceEpoch();
 
    Os::Logger::instance().log(FAC_SIP, PRI_DEBUG,
                  "SipPersistentSubscriptionMgr::updateVersion "
